requirements.txt
numpy
pandas
scikit-learn
tensorflow
flask
gunicorn
pillow
matplotlib
seaborn 
streamlit